refactor(util): expose `defineOperation` for testing
